refactor: divide master and worker codes
