+++ badjs-acceptor/service/ProjectService.js
@@ -3,6 +3,11 @@
  * Created by chriscai on 2015/1/23.
  */
 
+/**
+ * TODO: project.db 体积会随着项目的增加逐渐膨胀，每次项目更新后全量持久化到文件 I/O 时间会线性增长；并且每次启动后全量加载项目数据到内存比较粗暴；
+ * 可以考虑做 redis 缓存
+ */
+
 const fs = require('fs');
 const express = require('express');
 const bodyParser = require('body-parser');
@@ -20,41 +25,41 @@ if (!fs.existsSync(dbPath)) {
 const app = express();
 
 app.use(bodyParser.json({ limit: '10mb' }));
-app.use(bodyParser.urlencoded({
-    extended: true,
-    limit: 10 * 1024 * 1024
-}));
+app.use(
+    bodyParser.urlencoded({
+        extended: true,
+        limit: 10 * 1024 * 1024
+    })
+);
 
-const ProjectService = function (clusters) {
-    const dispatchCluster = function (data) {
+const ProjectService = function(clusters) {
+    const dispatchCluster = function(data) {
         for (var i = 0; i < clusters.length; i++) {
             clusters[i].send(data);
         }
     };
 
-    app.use('/getProjects', function (req, res) {
+    // 主进程接收 projects 更新，然后通知 woker 进程更新
+    app.use('/getProjects', function(req, res) {
         var param = req.query;
-        if (req.method === "POST") {
+        if (req.method === 'POST') {
             param = req.body;
         }
 
         if (param.auth != 'badjsAccepter' || !param.projectsInfo) {
-
         } else {
             dispatchCluster({
                 projectsInfo: param.projectsInfo
             });
 
-            fs.writeFile(dbPath, param.projectsInfo || "", function () {
+            fs.writeFile(dbPath, param.projectsInfo || '', function() {
                 logger.info('update project.db');
             });
         }
 
         res.writeHead(200);
         res.end();
-
-    })
-        .listen(9001);
+    }).listen(9001);
 
     var info = fs.readFileSync(dbPath, 'utf-8');
 
@@ -62,6 +67,5 @@ const ProjectService = function (clusters) {
         projectsInfo: info
     });
 };
-
 
 module.exports = ProjectService;
